Extract cmd functions
